Move BaseObserver back to common
