+++ packages/capacitor/src/adapter/CapacitorSQLiteAdapter.ts
@@ -2,6 +2,7 @@ import { CapacitorSQLite, SQLiteConnection, SQLiteDBConnection } from '@capacito
 import { Capacitor } from '@capacitor/core';
 
 import {
+  BaseObserver,
   BatchedUpdateNotification,
   ConnectionPool,
   DBAdapter,
@@ -11,7 +12,7 @@ import {
   LockContext,
   QueryResult
 } from '@powersync/web';
-import { BaseObserver, Mutex, timeoutSignal } from '@powersync/shared-internals';
+import { Mutex, timeoutSignal } from '@powersync/shared-internals';
 import { PowerSyncCore } from '../plugin/PowerSyncCore.js';
 import { messageForErrorCode } from '../plugin/PowerSyncPlugin.js';
 import { CapacitorSQLiteOpenFactoryOptions, DEFAULT_SQLITE_OPTIONS } from './CapacitorSQLiteOpenFactory.js';
