+++ packages/common/package.json
@@ -7,6 +7,7 @@
   },
   "description": "API definitions for PowerSync",
   "type": "module",
+  "types": "./lib/index.d.ts",
   "exports": {
     ".": {
       "types": "./lib/index.d.ts",
