+++ packages/node/src/db/WorkerConnectionPool.ts
@@ -4,6 +4,7 @@ import * as path from 'node:path';
 import { Worker } from 'node:worker_threads';
 
 import {
+  BaseObserver,
   BatchedUpdateNotification,
   ConnectionPool,
   DBAdapterDefaultMixin,
@@ -13,7 +14,7 @@ import {
   QueryResult,
   Transaction
 } from '@powersync/common';
-import { BaseObserver, Semaphore, timeoutSignal } from '@powersync/shared-internals';
+import { Semaphore, timeoutSignal } from '@powersync/shared-internals';
 import { Remote } from 'comlink';
 import { AsyncDatabase, AsyncDatabaseOpener } from './AsyncDatabase.js';
 import { RemoteConnection } from './RemoteConnection.js';
