+++ packages/react-native/src/db/adapters/react-native-quick-sqlite/RNQSDBAdapter.ts
@@ -1,4 +1,5 @@
 import {
+  BaseObserver,
   DBAdapter,
   DBAdapterListener,
   LockContext as PowerSyncLockContext,
@@ -11,7 +12,6 @@ import {
 } from '@powersync/common';
 import type { QuickSQLiteConnection, LockContext as RNQSLockContext } from '@journeyapps/react-native-quick-sqlite';
 import { QueryResult, SqlExecutor } from '@powersync/common';
-import { BaseObserver } from '@powersync/shared-internals';
 
 class RNQSConnectionPool extends BaseObserver<DBAdapterListener> implements ConnectionPool {
   constructor(
