+++ packages/web/src/db/adapters/wa-sqlite/DatabaseClient.ts
@@ -1,4 +1,5 @@
 import {
+  BaseObserver,
   QueryResult,
   LockContext,
   DBLockOptions,
@@ -14,7 +15,7 @@ import { ClientConnectionView } from './DatabaseServer.js';
 import { RawQueryResult } from './RawSqliteConnection.js';
 import * as Comlink from 'comlink';
 import type { ConnectToMultiDatabaseServerOptions } from '../../../worker/db/MultiDatabaseServer.js';
-import { BaseObserver, ConnectionClosedError } from '@powersync/shared-internals';
+import { ConnectionClosedError } from '@powersync/shared-internals';
 
 export interface OpenWorkerConnection {
   connect(config: ConnectToMultiDatabaseServerOptions): Promise<ClientConnectionView>;
